W-055 - create separate files for user's basic info and additional info
