+++ backend/plortal/src/main/java/com/plortal/plortal/controller/UserController.java
@@ -102,52 +102,4 @@ public class UserController {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal server error");
         }
     }
-
-    @PostMapping("/update-basic-info/{userId}")
-    public ResponseEntity<?> updateBasicInfo(@PathVariable Long userId, @RequestParam String basicInfo) {
-        try {
-            userService.updateBasicInfo(userId, basicInfo);
-            return ResponseEntity.ok("Basic info updated successfully");
-        } catch (UserNotExistException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal server error");
-        }
-    }
-
-    @PostMapping("/update-additional-info/{userId}")
-    public ResponseEntity<?> updateAdditionalInfo(@PathVariable Long userId, @RequestParam String additionalInfo) {
-        try {
-            userService.updateAdditionalInfo(userId, additionalInfo);
-            return ResponseEntity.ok("Additional info updated successfully");
-        } catch (UserNotExistException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal server error");
-        }
-    }
-
-    @GetMapping("/{userId}/basic-info")
-    public ResponseEntity<?> getBasicInfo(@PathVariable Long userId) {
-        try {
-            String basicInfo = userService.getBasicInfo(userId);
-            return ResponseEntity.ok(basicInfo);
-        } catch (UserNotExistException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal server error");
-        }
-    }
-
-    @GetMapping("/{userId}/additional-info")
-    public ResponseEntity<?> getAdditionalInfo(@PathVariable Long userId) {
-        try {
-            String additionalInfo = userService.getAdditionalInfo(userId);
-            return ResponseEntity.ok(additionalInfo);
-        } catch (UserNotExistException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal server error");
-        }
-    }
 }
