+++ backend/plortal/src/main/java/com/plortal/plortal/service/impl/UserServiceImpl.java
@@ -113,46 +113,4 @@ public class UserServiceImpl implements UserService {
             throw new UserNotExistException();
         }
     }
-
-    @Override
-    public void updateBasicInfo(Long userId, String basicInfo) {
-        Optional<User> userOptional = userRepository.findById(userId);
-        if (userOptional.isPresent()) {
-            User user = userOptional.get();
-            user.setBasicInfo(basicInfo);
-            userRepository.save(user);
-        } else {
-            throw new UserNotExistException();
-        }
-    }
-
-    @Override
-    public void updateAdditionalInfo(Long userId, String additionalInfo) {
-        Optional<User> userOptional = userRepository.findById(userId);
-        if (userOptional.isPresent()) {
-            User user = userOptional.get();
-            user.setAdditionalInfo(additionalInfo);
-            userRepository.save(user);
-        } else {
-            throw new UserNotExistException();
-        }
-    }
-
-    @Override
-    public String getBasicInfo(Long userId) throws UserNotExistException {
-        Optional<User> user = userRepository.findById(userId);
-        if (user.isEmpty()) {
-            throw new UserNotExistException();
-        }
-        return user.get().getBasicInfo();
-    }
-
-    @Override
-    public String getAdditionalInfo(Long userId) throws UserNotExistException {
-        Optional<User> user = userRepository.findById(userId);
-        if (user.isEmpty()) {
-            throw new UserNotExistException();
-        }
-        return user.get().getAdditionalInfo();
-    }
 }
